# swagger/swagger-spec.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/portfolios/{id}": {
			"get": {
				"operationId": "PortfolioController_getOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
...
	},
	"info": {
		"title": "Buy The Dip API Docs",
		"description": "Buy The Dip API Swagger 입니다.\nAuthorization이 필요없습니다.",
		"version": "1.0.0",
		"contact": {}
	},
...
	"servers": [
		{
			"url": "http://localhost:4000",
			"description": "local"
		}
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http",
				"in": "header",
				"name": "Authorization"
			}
		},
...
	}
}